inizio implementazione controller per i bonifici, fix exception handling

+++ common-test-fabrick/src/main/java/it/fabrick/test/dto/BalanceDTO.java
@@ -13,7 +13,7 @@ public class BalanceDTO implements Serializable {
 	private static final long serialVersionUID = -3949766244099251596L;
 	@CustomDate
 	private Date date;
-	private Float balance;
-	private Float availableBalance;
+	private Double balance;
+	private Double availableBalance;
 	private CurrencyEnum currency;
 }
